requirements.txt
pymysql
boto3
python-dotenv


